pyproject.toml
[tool.poetry]
name = "chatpdf"
version = "0.1.0"
description = "Chat with your PDFs from the command line."
authors = ["Jeff Coggshall <thenextlocalminima@gmail.com>"]
readme = "README.md"
packages = [
    { include = "chatpdf" },
    { include = "scripts", from = "src" },
    { include = "utils", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
anthropic = {version = "^0.42.0", extras = ["vertex"]}
beautifulsoup4 = "*"
bitsandbytes = "^0.42.0"
click = "*"
mistralai = "*"
markdown = "*"
requests = "*"
feedparser = "*"
google-genai = ">=1.10.0,<2.0.0"
instructor = "1.4.3"
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
jupyter = "^1.0.0"
matplotlib = "^3.9.0"
notebook = "^7.2.0"
nbconvert = "^7.13.1"
nltk = "^3.8.1"
numpy = "^2.1.0"
pandas = "^2.2.0"
pypdf = {version = ">=5.4.0,<6.0.0", extras = ["crypto"]}
pymupdf = "*"
pdfplumber = "*"
sentence-transformers = "*"
scikit-learn = "^1.5.0"
tensorboard = "^2.19.0"
standard-imghdr = "^3.13.0"
prompt-toolkit = "*"
rich = "*"
tabulate = "*"
sqlite-utils = "^3.36"

python-dotenv = "^1.0.0"
db-dtypes = "1.2.0"
ratelimit = "2.2.0"
regex = "^2023.10.3"
seaborn = "^0.13.0"
weasyprint = "*"
openreview-py = "*"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

# Add this new section to exclude specific files or directories
exclude = [
    "__pypackages__",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D"  # pydocstyle
]
ignore = [
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "D401",  # imperative mood
    "D103",
    "D104",
    "D100",
    "D101",
    "N806",
    "N803",
    "E501",
    "D107",
    "D102",
    "D201",
    "D202",
    "D205",
    "B904",
    "B008",
    "E402",
]

[tool.ruff.per-file-ignores]
"**/*.ipynb" = ["E402"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    # at least this three should be fine in tests:
    "S101", # asserts allowed in tests...
    "E712", # == for asserts
    "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
    # The below are debateable
    "PLR2004", # Magic value used in comparison, ...
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
]


[tool.poetry.scripts]
chatpdf = "chatpdf.cli:main"